MessageEncoder and MessageWriter
Gets rid of PDataWriter.  Turns out it's a bad idea to abstract the
presentation data pdus as a stream to the upper layers because they have
to care about keeping PDVs shorter than PDUs.  The reader implementation
will actually parse illegal stuff because it doesn't care.

Also, since we have to calculate the length of PDUs and PDVs before
writing them out, we have to buffer partial data, and if we have more
layers, we have more buffering, which leads to inefficient copying.