intentionally don't check an error
Testing code review tools.  Because the tested function has two return
values and we need the other one, this _s the error, indicating that
we're explicitly not checking.  Next test will be to do this for a
function that only returns an error so that we can implicitly ignore it.